CLN: Rename gridprop_import_eclrun
