+++ src/xtgeo/grid3d/_find_gridprop_in_eclrun.py
